% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rREML.R
\name{bcrREML}
\alias{bcrREML}
\title{Bias-corrected Robust REML Fit}
\usage{
bcrREML(
  formula,
  data,
  vi,
  params,
  k1 = 4.685,
  k2 = 5.12,
  maxit = 500,
  tol = .Machine$double.eps^0.5
)
}
\arguments{
\item{formula}{an object of class "\link[stats:formula]{formula}" (or one that can be coerced to that class):
a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{vi}{an optional vector of sampling variances to be used in the fitting process.}

\item{params}{an optional vector of initial guess for the parameters in the model. By default the results of "\link[lme4:lmer]{lmer}" will be used.}

\item{k1}{a positive numeric indicates the tuning parameter of bisquare loss for location estimation.
Defaults to 4.685 corresponding to 95\% asymptotic efficiency in the simple location problem.}

\item{k2}{a positive numeric indicates the tuning parameter of bisquare loss for scale estimation.
Defaults to 5.12 corresponding to 90\% asymptotic efficiency in the simple scale problem.}

\item{maxit}{maximum number of iterations. Defaults to 500.}

\item{tol}{a numeric indicates convergence tolerance.}
}
\value{
A list with components:
\itemize{
\item \code{coefTable} a data frame contains estimation, standard error, z statistic and p-value of Wald-type test.
\item \code{varcov} the variance-covariance matrix for the coefficients.
\item \code{s1} the correction factor for sampling variance.
\item \code{s2} the variance of the random intercept in 2-level meta-analysis.
\item \code{Std.resid} standardized residuals of the fitted model.
\item \code{convergence} 0 indicates the algorithm converged; 1 indicates the algorithm failed to converge.
}
}
\description{
Bias-corrected robust REML estimation for two-level linear mixed models (only support random intercept).
}
\examples{
fit <- bcrREML(smd_enemy ~ 1 + (1|Code), trophic, trophic$vi_enemy)
fit$coefTable
}
