% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rREML.R
\name{coefs}
\alias{coefs}
\title{Extract Fixed Coefficients}
\usage{
coefs(x)
}
\arguments{
\item{x}{a \code{rREML} object.}
}
\description{
A function for extracting fixed coefficients in the model.
}
\examples{
x <- rREML(smd_enemy ~ 1 + (1 | Code), trophic, trophic$vi_enemy)
coefs(x)

}
