% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rREML.R
\name{covs}
\alias{covs}
\title{Extract Variance-Covariance Matrix}
\usage{
covs(x)
}
\arguments{
\item{x}{a \code{rREML} object.}
}
\description{
A function for extracting variance-covariance matrix for fixed coefficients in the model.
}
\examples{
x <- rREML(smd_enemy ~ 1 + (1 | Code), trophic, trophic$vi_enemy)
covs(x)

}
