% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbrobm.R
\name{plot.MedRes}
\alias{plot.MedRes}
\title{DAG Plot}
\usage{
\method{plot}{MedRes}(
  x,
  ...,
  main = "Diagnosis Plot",
  xlab = "Theoretical Quantiles",
  ylab = "Sample Quantiles"
)
}
\arguments{
\item{x}{a \code{MedRes} object to show.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function to generate the DAG plot for a \code{MedRes} object.
}
