% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rREML.R
\name{plot.rREML}
\alias{plot.rREML}
\title{Diagnosis Plot}
\usage{
\method{plot}{rREML}(
  x,
  ...,
  main = "Diagnosis Plot",
  xlab = "Theoretical Quantiles",
  ylab = "Sample Quantiles"
)
}
\arguments{
\item{x}{a \code{rREML} object to show.}

\item{...}{further arguments passed to or from other methods.}

\item{main}{the main title for the plot.}

\item{xlab}{the label for x axis.}

\item{ylab}{the label for y axis.}
}
\description{
Generic function to generate the diagnosis plot for a \code{rREML} object.
}
\examples{
x <- rREML(smd_enemy ~ 1 + (1 | Code), trophic, trophic$vi_enemy)
plot(x)

}
