% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rREML.R
\name{pred}
\alias{pred}
\title{Model Predictions}
\usage{
pred(x, newdata)
}
\arguments{
\item{x}{a \code{rREML} object.}

\item{newdata}{a data frame contains same variables as fitted object x.}
}
\description{
A generic function for predicting within-level value of response.
}
\examples{
x <- rREML(smd_enemy ~ 1 + (1 | Code), trophic, trophic$vi_enemy)
pred(x, trophic)

}
