% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbrobm.R
\name{print.MedRes}
\alias{print.MedRes}
\title{Print MedRes Object}
\usage{
\method{print}{MedRes}(x)
}
\arguments{
\item{x}{a \code{MedRes} object to print.}
}
\description{
Generic function to print a \code{MedRes} object.
}
