% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rREML.R
\name{print.rREML}
\alias{print.rREML}
\title{Print rREML Object}
\usage{
\method{print}{rREML}(x, ...)
}
\arguments{
\item{x}{a \code{rREML} object to print.}
}
\description{
Generic function to print a \code{rREML} object.
}
\examples{
x <- rREML(smd_enemy ~ 1 + (1 | Code), trophic, trophic$vi_enemy)
print(x)

}
