% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbrobm.R
\name{showDefault.MedRes}
\alias{showDefault.MedRes}
\title{Show MedRes Object}
\usage{
showDefault.MedRes(x, ...)
}
\arguments{
\item{x}{a \code{MedRes} object to show.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function to show a \code{MedRes} object.
}
