% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rREML.R
\name{showDefault.rREML}
\alias{showDefault.rREML}
\title{Show rREML Object}
\usage{
showDefault.rREML(x, ...)
}
\arguments{
\item{x}{a \code{rREML} object to show.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function to show a \code{rREML} object.
}
\examples{
rREML(smd_enemy ~ 1 + (1 | Code), trophic, trophic$vi_enemy)

}
