% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbrobm.R
\name{vbrobm}
\alias{vbrobm}
\title{Variational Bayesian Robust Mediation Analysis}
\usage{
vbrobm(..., n = 1000, q = c(0.025, 0.975))
}
\arguments{
\item{...}{one or more \code{rREML} objects}

\item{n}{bootstrap replication times}

\item{q}{desired quantile for confidence interval}
}
\value{
A list with components:
\itemize{
\item \code{coefs} a data frame contains estimation, standard error, z statistic and p-value of Wald-type test for all paths.
\item \code{mediations} a summary for direct, indirect and total effects of the fitted models.
}
}
\description{
Overall variational Bayesian robust mediation analysis for a direct acyclic graph.
}
\examples{
res <- vbrobm(vbrob(smd_enemy~1+(1|Code), data = trophic, vi = trophic$vi_enemy, maxit = 500),
              vbrob(smd_herbivore~1+smd_enemy+(1|Code), data = trophic, vi = trophic$vi_herbivore, maxit = 500),
              vbrob(smd_plant~1+smd_herbivore+(1|Code), data = trophic, vi = trophic$vi_plant, maxit = 500))
res
plot(res)
}
